Add files via upload
This is my first trial training my MAMA based on 8000 dataset i was able to gather. After different hyperparameter tuning there were no increased sign of generalization. It kept memorizing and overfitting. This could be as a result of my small dataset. So i am going to do Transfer Learning after this.