Good Validation accuracy but poor intent classification from input
My model had a validation accuracy of 91% which i believed is actually a very good score for an NLP project. To my biggest surprise, the model performed woefully on my input test. So i am going to do further research and see if the Dataset could actually be the cause for this.